bilbo and Mummi contribution!! veriverse contribution!! LCTV contribution!! catherinacortega (helenawinter), davidwisser, drmjg, frankly444, fuel, david_joffe, dopedopedope, ggc (Davit), intelijens, jeffersonciotti, kaeris,  pablosj (Pawel Kulesza), r00tkiller, wapaca and wiredoncoffee (from North Calgary , Alberta) contribuntion!! Twitch contribution!! kdex___, keptar and sonnystartminecraft contribution!! Beampro contribution!!
From bilbo-hyperv
i7 6700

+++ appseed/a_spalib/main.cpp
@@ -264,6 +264,15 @@ int32_t a_spa::run()
 int a_spa::spa_main()
 {
 
+   spa_mutex mutex;
+
+   if (mutex.already_exists())
+   {
+
+      return -34;
+
+   }
+
    string str(::GetCommandLineW());
 
    if(str.find(" install ") < 0)
